refactor:repository 도메인별 폴더 정리
