+++ src/main/java/bluepill/server/repository/character/CharacterCardRepositoryCustom.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.character;
 
 import bluepill.server.dto.character.CharacterCardListItem;
 import bluepill.server.dto.character.CharacterSortType;
