+++ src/main/java/bluepill/server/repository/chat/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.chat;
 
 import bluepill.server.domain.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
