+++ src/main/java/bluepill/server/repository/chat/ChatMessageRepositoryCustom.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.chat;
 
 import bluepill.server.domain.ChatMessage;
