+++ src/main/java/bluepill/server/repository/logroom/CharacterPhotoRow.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.logroom;
 
 import java.time.LocalDate;
