+++ src/main/java/bluepill/server/repository/logroom/LogPhotoRepositoryCustom.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.logroom;
 
 import java.time.LocalDate;
 import java.util.List;
