+++ src/main/java/bluepill/server/repository/logroom/LogRoomMemberRepository.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.logroom;
 
 import bluepill.server.domain.LogRoomMember;
 import org.springframework.data.jpa.repository.JpaRepository;
