+++ src/main/java/bluepill/server/repository/logroom/LogRoomRelationshipRepository.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.logroom;
 
 import bluepill.server.domain.LogRoomRelationship;
 import org.springframework.data.jpa.repository.JpaRepository;
