+++ src/main/java/bluepill/server/repository/logroom/LogRoomRepositoryCustom.java
@@ -1,4 +1,4 @@
-package bluepill.server.repository;
+package bluepill.server.repository.logroom;
 
 import java.util.List;
 import java.util.UUID;
