+++ src/main/java/bluepill/server/scheduler/DailyLogScheduler.java
@@ -2,7 +2,7 @@ package bluepill.server.scheduler;
 
 import bluepill.server.client.AgentClient;
 import bluepill.server.client.AgentClient.AgentDailyLogRequest;
-import bluepill.server.repository.LogRoomMemberRepository;
+import bluepill.server.repository.logroom.LogRoomMemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
