+++ src/main/java/bluepill/server/service/AuthService.java
@@ -6,8 +6,8 @@ import bluepill.server.dto.user.TokenResponse;
 import bluepill.server.exception.BusinessException;
 import bluepill.server.exception.ErrorCode;
 import bluepill.server.jwt.JwtProvider;
-import bluepill.server.repository.UserRepository;
-import bluepill.server.repository.UserTokenRepository;
+import bluepill.server.repository.user.UserRepository;
+import bluepill.server.repository.user.UserTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
