+++ src/main/java/bluepill/server/service/ChatAiReplyService.java
@@ -5,10 +5,10 @@ import bluepill.server.domain.CharacterSnapshot;
 import bluepill.server.domain.ChatMessage;
 import bluepill.server.domain.LogRoom;
 import bluepill.server.dto.chat.ChatMessageItem;
-import bluepill.server.repository.ChatMessageRepository;
-import bluepill.server.repository.ChatSseEmitterRepository;
-import bluepill.server.repository.LogRoomMemberRepository;
-import bluepill.server.repository.LogRoomRepository;
+import bluepill.server.repository.chat.ChatMessageRepository;
+import bluepill.server.repository.chat.ChatSseEmitterRepository;
+import bluepill.server.repository.logroom.LogRoomMemberRepository;
+import bluepill.server.repository.logroom.LogRoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
