+++ src/main/java/bluepill/server/service/OAuth2SuccessHandler.java
@@ -6,8 +6,8 @@ import bluepill.server.exception.BusinessException;
 import bluepill.server.exception.ErrorCode;
 import bluepill.server.jwt.JwtConfig;
 import bluepill.server.jwt.JwtProvider;
-import bluepill.server.repository.UserRepository;
-import bluepill.server.repository.UserTokenRepository;
+import bluepill.server.repository.user.UserRepository;
+import bluepill.server.repository.user.UserTokenRepository;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
