+++ src/main/java/bluepill/server/service/OAuth2UserService.java
@@ -4,8 +4,8 @@ import bluepill.server.domain.SubscriptionPlan;
 import bluepill.server.domain.User;
 import bluepill.server.exception.BusinessException;
 import bluepill.server.exception.ErrorCode;
-import bluepill.server.repository.SubscriptionPlanRepository;
-import bluepill.server.repository.UserRepository;
+import bluepill.server.repository.user.SubscriptionPlanRepository;
+import bluepill.server.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
