+++ src/main/java/bluepill/server/service/PostService.java
@@ -12,12 +12,12 @@ import bluepill.server.dto.post.PostShareResponse;
 import bluepill.server.dto.post.PostSharer;
 import bluepill.server.exception.BusinessException;
 import bluepill.server.exception.ErrorCode;
-import bluepill.server.repository.LogRoomMemberRepository;
-import bluepill.server.repository.LogRoomRepository;
-import bluepill.server.repository.MemberImageRow;
-import bluepill.server.repository.PostPageRow;
-import bluepill.server.repository.PostPhotoRow;
-import bluepill.server.repository.PostRepository;
+import bluepill.server.repository.logroom.LogRoomMemberRepository;
+import bluepill.server.repository.logroom.LogRoomRepository;
+import bluepill.server.repository.logroom.MemberImageRow;
+import bluepill.server.repository.post.PostPageRow;
+import bluepill.server.repository.post.PostPhotoRow;
+import bluepill.server.repository.post.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
