+++ src/main/java/bluepill/server/service/UserService.java
@@ -6,9 +6,9 @@ import bluepill.server.dto.user.UserProfileResponse;
 import bluepill.server.dto.user.UserProfileUpdateRequest;
 import bluepill.server.exception.BusinessException;
 import bluepill.server.exception.ErrorCode;
-import bluepill.server.repository.CharacterCardRepository;
-import bluepill.server.repository.PostRepository;
-import bluepill.server.repository.UserRepository;
+import bluepill.server.repository.character.CharacterCardRepository;
+import bluepill.server.repository.post.PostRepository;
+import bluepill.server.repository.user.UserRepository;
 import bluepill.server.util.ImageUrlBuilder;
 import bluepill.server.util.NicknameGenerator;
 import lombok.RequiredArgsConstructor;
